+++ CHANGELOG.md
@@ -15,7 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Treasury page shell
 - Bounties page shell
 - Tech comm page shell
+- Tooltip directive
 - Params info on recent events
+
+#### Removed
+- Tooltip component
 
 ## [0.1.0-rc.2] - 2021-10-15
 #### Added