+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0-rc.3] - 2021-10-19
 #### Added
 - Auctions tab shell
 - Crowdloan tab shell
@@ -109,6 +109,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Initial project release
 - Initiated project with vue cli
 
-[Unreleased]: https://gitlab.com/tokend/polkadot/new-web-client/compare/0.1.0-rc.2...main
+[Unreleased]: https://gitlab.com/tokend/polkadot/new-web-client/compare/0.1.0-rc.3...HEAD
+[0.1.0-rc.3]: https://gitlab.com/tokend/polkadot/new-web-client/compare/0.1.0-rc.2...0.1.0-rc.3
 [0.1.0-rc.2]: https://gitlab.com/tokend/polkadot/new-web-client/compare/0.1.0-rc.1...0.1.0-rc.2
 [0.1.0-rc.1]: https://gitlab.com/tokend/polkadot/new-web-client/tags/0.1.0-rc.1